requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
auth==0.5.3
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
eventlet==0.36.1
falcon==3.1.3
fastapi==0.110.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
jose==1.0.0
Mako==1.3.2
MarkupSafe==2.1.5
mongoengine==0.28.2
outcome==1.3.0.post0
packaging==24.0
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
starlette==0.37.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
wsproto==1.2.0
